Some code fixes
