+++ Assets/Scenes/Scripts/Enemies/ParticlesDestroy.cs
