+++ Assets/Scenes/Scripts/NPCs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 04da3f99a86df7b49a018a32e60e425d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
