+++ Assets/Scenes/Scripts/NPCs/NPC.cs.meta
