+++ Assets/Scenes/Scripts/NPCs/scriptableNPC.cs
