+++ Assets/Scenes/Scripts/Players/PlayerArcher.cs
@@ -4,21 +4,17 @@ using UnityEngine;
 
 public class PlayerArcher : Heroes
 {
-    Animator archerAnimator;
     public Vector2 Mov { get { return mov; } }
 
     public int movX;
     public int movY;
 
-    // Start is called before the first frame update
     void Start()
     {
         Anim = GetComponent<Animator>();
         Rb = GetComponent<Rigidbody2D>();
-
     }
 
-    // Update is called once per frame
     void Update()
     {
         mov = new Vector2(Input.GetAxisRaw("Horizontal_P2"), Input.GetAxisRaw("Vertical_P2"));
@@ -30,7 +26,6 @@ public class PlayerArcher : Heroes
     private void FixedUpdate()
     {
         Movement();
-
         Die();
     }
 
@@ -38,7 +33,7 @@ public class PlayerArcher : Heroes
     {
         if (collision.gameObject.CompareTag("Enemy"))
         {
-            lifeQuantity--;
+            hp--;
         }
     }
 }
