+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.3.19f1
-m_EditorVersionWithRevision: 2021.3.19f1 (c9714fde33b6)
+m_EditorVersion: 2021.3.10f1
+m_EditorVersionWithRevision: 2021.3.10f1 (1c7d0df0160b)
